# Use instance pool allocation strategy for wasmtime shim