--- a/Cargo.toml
+++ b/Cargo.toml
@@ -52,9 +52,9 @@ tokio = { version = "1.41.1", default-features = false }
 tokio-util = { version = "0.7", default-features = false }
 
 # wasmtime
-wasmtime = { version = "26.0.1", features = ["async"] }
-wasmtime-wasi = { version = "26.0.1" }
-wasmtime-wasi-http = { version = "26.0.1" }
+wasmtime = { version = "27.0.0", features = ["async"] }
+wasmtime-wasi = { version = "27.0.0" }
+wasmtime-wasi-http = { version = "27.0.0" }
 
 [profile.release]
 panic = "abort"