--- a/benches/containerd-shim-benchmarks/benches/wasmedge-benchmarks.rs
+++ b/benches/containerd-shim-benchmarks/benches/wasmedge-benchmarks.rs
@@ -151,4 +151,3 @@ criterion_group! {
 }
 
 criterion_main!(benches);
-