This patch changes `wasm` instance allocation strategy from `OnDemand` to `PoolAllocator` which increase a performance of creating a per request instances for components targeting `wasi/http` world.

Here, before and after throughput, for a minimum handler that responds with `Hello World`, shows 10x improvement:

Before:
```
hey -n 10000 -c 10 http://127.0.0.1:8080

Summary:
  Total:        2.6978 secs
  Slowest:      0.0135 secs
  Fastest:      0.0007 secs
  Average:      0.0027 secs
  Requests/sec: 3706.6977


Response time histogram:
  0.001 [1]     |
  0.002 [1323]  |■■■■■■■■
  0.003 [6765]  |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.005 [1769]  |■■■■■■■■■■
  0.006 [123]   |■
  0.007 [7]     |
  0.008 [1]     |
  0.010 [4]     |
  0.011 [4]     |
  0.012 [2]     |
  0.013 [1]     |


Latency distribution:
  10% in 0.0019 secs
  25% in 0.0022 secs
  50% in 0.0026 secs
  75% in 0.0031 secs
  90% in 0.0036 secs
  95% in 0.0039 secs
  99% in 0.0047 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0000 secs, 0.0007 secs, 0.0135 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0000 secs
  req write:    0.0000 secs, 0.0000 secs, 0.0005 secs
  resp wait:    0.0026 secs, 0.0006 secs, 0.0123 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.0009 secs

Status code distribution:
  [200] 10000 responses
```

After:

```
hey -n 10000 -c 10 http://127.0.0.1:8080

Summary:
  Total:        0.2566 secs
  Slowest:      0.0017 secs
  Fastest:      0.0001 secs
  Average:      0.0003 secs
  Requests/sec: 38975.5182
  

Response time histogram:
  0.000 [1]     |
  0.000 [5736]  |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.000 [3507]  |■■■■■■■■■■■■■■■■■■■■■■■■
  0.001 [478]   |■■■
  0.001 [159]   |■
  0.001 [72]    |■
  0.001 [28]    |
  0.001 [10]    |
  0.001 [4]     |
  0.002 [2]     |
  0.002 [3]     |


Latency distribution:
  10% in 0.0001 secs
  25% in 0.0002 secs
  50% in 0.0002 secs
  75% in 0.0003 secs
  90% in 0.0004 secs
  95% in 0.0005 secs
  99% in 0.0008 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0000 secs, 0.0001 secs, 0.0017 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0000 secs
  req write:    0.0000 secs, 0.0000 secs, 0.0014 secs
  resp wait:    0.0002 secs, 0.0001 secs, 0.0012 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.0009 secs

Status code distribution:
  [200] 10000 responses
```

Closes: https://github.com/containerd/runwasi/issues/729